requirements.txt
tensorflow>=2.13.0
keras-cv>=0.6.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
safetensors>=0.3.0
pillow>=9.0.0
matplotlib>=3.6.0
numpy>=1.24.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0
opencv-python>=4.8.0
controlnet-aux>=0.0.6
imageio[ffmpeg]>=2.31.0
streamlit>=1.25.0
invisible-watermark>=0.2.0
xformers>=0.0.20